Improve tag filtering
Filtering by tags is now using AND instead of OR logic. That is:
Before: If you select tags mottagning and gasque, you would get patches
which have the tags mottagning OR gasque, i.e. you could get patches with tags
mottagning and finsittning and so on.
Now: If you select tags mottagning and gasque, you only get patches with that combination.